Rearrange component folders to work better with Firebase components. Add Pages folder to store the rendered pages. User state is now passed down via React Context.
