+++ package.json
@@ -16,7 +16,7 @@
     "react-router": "^5.2.1",
     "react-router-dom": "^5.3.0",
     "react-scripts": "4.0.3",
-    "react-select": "^4.3.1",
+    "react-select": "^5.1.0",
     "react-to-print": "^2.13.0",
     "sass": "^1.39.0",
     "scss": "^0.2.4"
