+++ src/App.css
@@ -4,67 +4,4 @@
     margin: 0;
 }
 
-.header {
-    background-color: black;
-}
 
-h1 {
-    font-size: 3rem;
-    text-align: center;
-}
-
-.main-title {
-    font-size: 6rem;
-   /* font-family: 'Franklin Gothic Medium', 'Arial Narrow', Arial, sans-serif;  */
-    font-family: Impact, Haettenschweiler, 'Arial Narrow Bold', sans-serif;
-    letter-spacing: 2rem;
-    /* text-shadow: 1px 1px bisque; */
-    text-shadow: 2px 2px royalblue;
-}
-
-input {
-    min-width: 300px;
-}
-
-.user-interface-window {
-    display: flex;
-    flex-direction: column;
-    background-color: black;
-    color: whitesmoke;
-    list-style: none;
-    align-items: center;
-}
-
-h2 {
-    font-size: 4rem;
-    text-align: center;
-    text-shadow: whitesmoke 5px 2px;
-}
-
-h3 {
-    font-size: 6rem;
-    text-align: center;
-    text-shadow: darkgray 3px 2px;
-}
-
-button:hover {
-    cursor: pointer;
-}
-
-.toolbar-container {
-    display: flex;
-    justify-content: center;
-    flex-direction: row;
-    background-color: black;
-}
-
-.tool-button {
-    background-color: black;
-    color: white;
-    padding: 2rem;
-    width: 10rem;
-    margin: 10%;
-    font-size: 1rem;
-    font-weight: bold;
-    border-radius: 10%;
-}
