+++ src/Components/Header/Header.component.jsx
@@ -1,37 +1,30 @@
 import { signOut } from "firebase/auth";
 import { auth } from "../Firebase/firebase.utils";
-import { BackFromSignIn, SignInButton } from "../SignInAndSignUp/signinButtons";
+import { BackFromSignIn, SignInButton } from "../SignInButtons/signinButtons";
 
 import './header.styles.scss';
 import { useEffect } from "react";
 import LoadCharacter from "../LoadCharacter/LoadCharacter";
 import { useLocation } from "react-router";
-
-
+import { useContext } from "react";
+import { UserContext } from "../../context";
 
 const Header = (props) => {
-    const currentUser = props.currentUser;
+    const { user } = useContext(UserContext);
+    
     const setSelectedAdvantagesList = props.setSelectedAdvantagesList;
     const setSelectedDisadvantagesList = props.setSelectedDisadvantagesList;
-
-    const selectedAdvantagesList = props.selectedAdvantagesList;
 
     const location = useLocation();
 
 
     const signout = () => {
         signOut(auth).then(() => {
-            console.log(currentUser + ' signed out');
+            console.log(user + ' signed out');
         }).catch((error) => {
             console.log('error signing out user', error.message);
         });
 
-    }
-
-    const con = () => {
-        console.log(currentUser);
-        console.log('location: ', location.pathname);
-        console.log('advantages list: ', selectedAdvantagesList);
     }
 
     useEffect(() => {
@@ -41,16 +34,15 @@ const Header = (props) => {
     return (
         <div className="header">
             <div className="sign-in-page">
-                {(location.pathname !== '/sign-in-and-sign-up' & !currentUser) &&
-                    <SignInButton currentUser={currentUser} />}
-                {currentUser && <button className="signout" onClick={() => signout()}></button>}
-                <button onClick={con}>con</button>
+                {(location.pathname !== '/sign-in-and-sign-up' & !user) &&
+                    <SignInButton />}
+                {user && <button className="signout" onClick={() => signout()}></button>}
             </div>
 
-            {currentUser &&
+            {user &&
                 <div className="load-character">
                     <LoadCharacter
-                        currentUser={currentUser}
+                        user={user}
                         setSelectedAdvantagesList={setSelectedAdvantagesList}
                         setSelectedDisadvantagesList={setSelectedDisadvantagesList}
                     />
