--- src/Components/SignInAndSignUp/SignIn/sign-in.styles.scss
@@ -1,65 +1,0 @@
-.sign-in {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    // width: 50vw;
-    padding: 5%;
-
-    background-color: black;
-    color: white;
-
-    span {
-        margin: 10% 0;
-    }
-
-    form {
-        display: flex;
-        flex-direction: column;
-
-
-        // width: 40%;
-
-        input {
-            margin: 5% 5%;
-            width: 80%;
-        }
-
-        label {
-            margin-top: 5%;
-        }
-
-        button {
-            min-width: 165px;
-            width: auto;
-            height: 50px;
-            letter-spacing: 0.5px;
-            line-height: 50px;
-            padding: 0 35px 0 35px;
-            font-size: 15px;
-            background-color: black;
-            color: white;
-            text-transform: uppercase;
-            font-family: 'Open Sans Condensed';
-            font-weight: bolder;
-            border: none;
-            cursor: pointer;
-
-            &:hover {
-                background-color: white;
-                color: black;
-                border: 1px solid black;
-            }
-
-            &.google-sign-in {
-                background-color: #4285f4;
-                color: white;
-
-                &:hover {
-                    background-color: #357ae8;
-                    border: none;
-                }
-            }
-        }
-    }
-}
