+++ src/Components/SignInButtons/signInButtons.styles.scss
@@ -1,28 +1,11 @@
 .sign-in-link {
     display: flex;
-    // justify-content: end;
     width: 100%;
     height: 100%;
     background-color: black;
    
- 
     a {
        margin: 1rem 2rem 0 auto;
        color: white;
     }
  }
-
-
-    
-
- 
-  
-     .forms {
-         display: flex;
-         justify-content: left;
-         height: 100vh;
-        
-        
-     }
-
-
