+++ src/Components/SignInButtons/signinButtons.jsx
@@ -1,9 +1,9 @@
 import { Link } from "react-router-dom";
 
-import './sign-in-and-sign-up.scss';
+import './signInButtons.styles.scss';
 
 export const SignInButton = () => {
- 
+
     return (
         <div className='sign-in-link'>
             <Link to='/sign-in-and-sign-up'>SIGN IN</Link>
@@ -14,7 +14,7 @@ export const SignInButton = () => {
 export const BackFromSignIn = () => {
     return (
         <div className="sign-in-link">
-        <Link to='/' >BACK</Link>
-    </div>
+            <Link to='/' >BACK</Link>
+        </div>
     )
 }
