+++ src/Components/saveCharacter/save-character-component.js
@@ -1,11 +1,14 @@
+import { useContext } from "react";
+import { UserContext } from "../../context";
 import { addNewCharacterForUser } from "../Firebase/firebase.utils";
 
 const SaveCharacter = (props) => {
+    const {user} = useContext(UserContext);
+
     const selectedAdvantagesList = props.selectedAdvantagesList;
     const selectedDisadvantagesList = props.selectedDisadvantagesList;
     const characterName = props.characterName;
 
-    const currentUser = props.currentUser;
     const saveCharacterHandler = async () => {
         console.log("****saveCharacterHandler Called");
         console.log("****selectedAdvantagesList: ", selectedAdvantagesList);
@@ -19,22 +22,15 @@ const SaveCharacter = (props) => {
             disadvantages: selectedDisadvantagesList.map(({ title }) => title)
         };
 
-        const currentlyLoggedInUserId = currentUser.uid;
+        const currentlyLoggedInUserId = user.uid;
         console.log("**** New Character for " + currentlyLoggedInUserId + " is ", newCharacter);
         const newCharacterRef = await addNewCharacterForUser(currentlyLoggedInUserId, newCharacter);
         console.log("**** newCharacterRef: ", newCharacterRef);
     };
 
-    const con = () => {
-        console.log('advantages list: ', selectedAdvantagesList);
-        console.log('disadvantages list: ', selectedDisadvantagesList);
-        console.log('name: ', characterName);
-    }
-
     return (
         <div className="button-container">
             <button className="tool-button" onClick={saveCharacterHandler}>Save Character</button>
-            <button onClick={con}>con</button>
         </div>
     );
 }
