+++ src/Pages/SignInAndSignUp/sign-in-and-sign-up.scss
@@ -1,0 +1,70 @@
+.forms {
+  display: flex;
+  justify-content: left;
+  height: 100vh;
+  width: 100%;
+}
+
+.sign-in, .sign-up {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  width: 50%;
+  background-color: black;
+  color: white;
+
+  span {
+      margin-top: 3%;
+      font-size: 1.5rem;
+  }
+
+  form {
+      display: flex;
+      flex-direction: column;
+      width: 60%;
+      padding: 5%;
+
+      input {
+          margin: 5% 5%;
+          width: 60%;
+      }
+
+      label {
+          margin: 5% 2% 0 0;
+      }
+
+      button {
+          min-width: 165px;
+          width: auto;
+          height: 50px;
+          letter-spacing: 0.5px;
+          line-height: 50px;
+          padding: 0 35px 0 35px;
+          font-size: 15px;
+          background-color: black;
+          color: white;
+          text-transform: uppercase;
+          font-family: 'Open Sans Condensed';
+          font-weight: bolder;
+          border: none;
+          cursor: pointer;
+
+          &:hover {
+              background-color: white;
+              color: black;
+              border: 1px solid black;
+          }
+
+          &.google-sign-in {
+              background-color: #4285f4;
+              color: white;
+
+              &:hover {
+                  background-color: #357ae8;
+                  border: none;
+              }
+          }
+      }
+  }
+}
